I hope it's good now

+++ functions_nested_loops/102-fibonacci.c
@@ -9,15 +9,15 @@
 int main(void)
 {
 	int n = 50;
-	unsigned int F1 = 1, F2 = 2, Fn;
+	unsigned long int F1 = 1, F2 = 2, Fn;
 	int i = 1;
-	printf("%u, %u", F1, F2);
+	printf("%lu, %lu", F1, F2);
 	for (i = 1 ; i <= n; i++)
 	{
 		if (i >= 3)
 		{
 		Fn = F1 + F2;
-		printf(", %u", Fn);
+		printf(", %lu", Fn);
 
 		F1 = F2;
 		F2 = Fn;
